src/sim/sg13g2/requirements.txt
pytest
cocotb
cocotbext-spi